Modified char * malloc size to have a x1.5 multiplier factor.
Improved all the logs at server side.
Fixed problem with throughput.

+++ include/mysnmp.h
@@ -44,7 +44,7 @@ static struct session {
 
 int snmpStop;
 char * snmpDestination;
-int ramSize;
+unsigned long ramSize;
 struct timeval tStart;
